+++ packages/app/src/handlers/index.ts
@@ -1,0 +1,3 @@
+export { setLogShipping } from './set-log-shipping';
+export { setRetention } from './set-retention';
+export { shipLogEntries } from './ship-log-entries';
