+++ packages/app/src/handlers/set-log-shipping.ts
@@ -1,16 +1,11 @@
 import { CloudWatchLogs, ECS, Lambda, S3 } from 'aws-sdk';
 import {
 	getCloudWatchLogGroups,
-	setCloudwatchRetention,
 	subscribeGroups,
 	unsubscribeGroups,
-} from './cloudwatch';
-import {
-	getCommonConfig,
-	getConfigureLogShippingConfig,
-	getSetRetentionConfig,
-} from './config';
-import { updateStructuredFieldsData } from './structuredFields';
+} from '../cloudwatch';
+import { getCommonConfig, getConfigureLogShippingConfig } from '../config';
+import { updateStructuredFieldsData } from '../structuredFields';
 
 const { awsConfig } = getCommonConfig();
 
@@ -18,40 +13,6 @@ const cloudwatchLogs = new CloudWatchLogs(awsConfig);
 const s3 = new S3(awsConfig);
 const lambda = new Lambda(awsConfig);
 const ecs = new ECS(awsConfig);
-
-function sleep(ms: number) {
-	return new Promise((resolve) => {
-		setTimeout(resolve, ms);
-	});
-}
-
-export async function setRetention(): Promise<void> {
-	const { retentionInDays } = getSetRetentionConfig();
-	const cloudwatchLogGroups = await getCloudWatchLogGroups(cloudwatchLogs);
-
-	for (const logGroup of cloudwatchLogGroups) {
-		if (logGroup.logGroupName === undefined) {
-			break; // cannot do anything
-		}
-
-		if (logGroup.retentionInDays === retentionInDays) {
-			console.log(
-				`Log group ${logGroup.logGroupName} retention is already ${retentionInDays} days`,
-			);
-		} else {
-			await setCloudwatchRetention(
-				cloudwatchLogs,
-				logGroup.logGroupName,
-				retentionInDays,
-			);
-			// avoid hitting the SDK throttling limit
-			await sleep(200);
-			console.log(
-				`Set ${logGroup.logGroupName} retention to ${retentionInDays} days`,
-			);
-		}
-	}
-}
 
 function eligibleForLogShipping(
 	logNamePrefixes: string[],
