+++ packages/app/src/handlers/ship-log-entries.ts
@@ -6,12 +6,12 @@ import type {
 } from 'aws-lambda';
 import { Kinesis, S3 } from 'aws-sdk';
 import type { PutRecordsOutput } from 'aws-sdk/clients/kinesis';
-import { BUILD_INFO } from './build-info';
-import { getCommonConfig, getShipLogsConfig } from './config';
-import { putKinesisRecords } from './kinesis';
-import { createStructuredLog } from './logEntryProcessing';
-import type { StructuredFields } from './model';
-import { getStructuredFields } from './structuredFields';
+import { BUILD_INFO } from '../build-info';
+import { getCommonConfig, getShipLogsConfig } from '../config';
+import { putKinesisRecords } from '../kinesis';
+import { createStructuredLog } from '../logEntryProcessing';
+import type { StructuredFields } from '../model';
+import { getStructuredFields } from '../structuredFields';
 
 const { awsConfig } = getCommonConfig();
 const { kinesisStreamName, structuredDataBucket, structuredDataKey } =
